ai-services/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
httpx==0.28.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0
